stop showing links for all page numbers so large numbers of pages don't overwhelm the user

+++ classes/model/kwalbum/item.php
@@ -1116,10 +1116,15 @@ class Model_Kwalbum_Item extends Kwalbum_Model
 		return Model::factory('kwalbum_item')->load((int)$result[0]['id']);
 	}
 
+    /**
+     * @param int $index
+     * @param boolean $in_edit_mode
+     * @return int
+     */
 	static public function get_page_number($index, $in_edit_mode = false)
 	{
         $multiplier = $in_edit_mode ? Model_Kwalbum_Item::EDIT_THUMB_MULTIPLIER : 1;
-        return ceil($index/(self::get_config('items_per_page')*$multiplier));
+        return (int) ceil($index / (self::get_config('items_per_page') * $multiplier));
 	}
 
 	/**
